+++ Basil/include/Renderer.h
@@ -3,16 +3,18 @@
  */
 #pragma once
 
+#include "OrthographicCamera.h"
 #include "RendererAPI.h"
+#include "Shader.h"
 
 namespace Basil
 {
 	class Renderer
 	{
 		public:
-			static void beginScene();
+			static void beginScene(OrthographicCamera& camera);
 			static void endScene();
-			static void submit(const std::shared_ptr<VertexArray>& vao);
+			static void submit(const std::shared_ptr<Shader>& shader, const std::shared_ptr<VertexArray>& vao);
 			
 			static void setClearColor(const glm::vec4& color);
 			static void clear();
@@ -20,6 +22,12 @@ namespace Basil
 
 			static RendererAPI::API getAPI();
 		private:
+			struct SceneData
+			{
+				glm::mat4 viewProjectionMatrix;
+			};
+
 			static RendererAPI* rendererAPI;
+			static SceneData* sceneData;
 	};
 }