+++ Basil/include/Shader.h
@@ -5,6 +5,8 @@
 
 #include <string>
 
+#include <glm/glm.hpp>
+
 namespace Basil
 {
 	class Shader
@@ -14,6 +16,7 @@ namespace Basil
 			~Shader();
 			void bind() const;
 			void unbind() const;
+			void uploadUniformMat4(const std::string& name, const glm::mat4& matrix);
 		private:
 			uint32_t rendererID;
 	};