+++ Basil/src/Shader.cpp
@@ -2,6 +2,7 @@
 #include "Shader.h"
 
 #include <glad/glad.h>
+#include <glm/gtc/type_ptr.hpp>
 
 namespace Basil
 {
@@ -116,5 +117,12 @@ namespace Basil
 	void Shader::unbind() const
 	{
 		glUseProgram(0);
+	}
+
+	// Upload a mat4 uniform
+	void Shader::uploadUniformMat4(const std::string& name, const glm::mat4& matrix)
+	{
+		GLint location = glGetUniformLocation(rendererID, name.c_str());
+		glUniformMatrix4fv(location, 1, GL_FALSE, glm::value_ptr(matrix));
 	}
 }